requirements.txt
metaapi-cloud-sdk>=29.1.1
pandas>=2.3.3
python-dotenv>=1.2.1
python-telegram-bot>=22.5
requests>=2.32.5
textblob>=0.19.0
yfinance>=0.2.66
flask>=3.0.0
matplotlib>=3.8.0
